backend/requirements.txt
flask==3.1.0
flask_cors==5.0.1
requests==2.31.0
spacy==3.8.5
sentence_transformers==4.1.0
protobuf==6.30.2
tiktoken==0.9.0
